x64 portability
git-svn-id: http://miranda.googlecode.com/svn/trunk@9182 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/include/m_protocols.h
@@ -289,7 +289,7 @@ typedef struct tagACCOUNT
 #define MS_PROTO_ENUMACCOUNTS "Proto/EnumAccounts"
 
 __inline static int ProtoEnumAccounts( int* accNumber, PROTOACCOUNT*** accArray )
-{	return CallService( MS_PROTO_ENUMACCOUNTS, ( WPARAM )accNumber, (LPARAM)accArray );
+{	return (int)CallService( MS_PROTO_ENUMACCOUNTS, ( WPARAM )accNumber, (LPARAM)accArray );
 }
 
 //retrieves an account's interface by its physical name (database module)
